pool: hoist ban/whitelist subsystem into pool::SharechainNode (kill node.cpp drift, slice 1)
node.cpp is copy-pasted across ltc/btc/bch/dgb. This introduces the
intermediate layer pool::SharechainNode<Config,ShareChain,Peer> between
pool::BaseNode and each coin's NodeImpl, and hoists the FIRST byte-identical,
coin-agnostic, non-consensus block into it: the peer ban-list + whitelist
subsystem.

Hoisted (verified byte-identical across all four coins, coin namespace aside):
  is_banned, is_whitelisted, set_ban_duration, set_whitelist_path,
  load_whitelist_from_disk, save_whitelist_to_disk,
  admin_list_bans, admin_ban_ip, admin_unban_ip, admin_list_whitelist,
  admin_whitelist_remove, and the build_bans_json/build_whitelist_json helpers,
  plus the members m_ban_list, m_ip_ban_list, m_ban_duration,
  m_whitelist_ips, m_whitelist_hosts, m_whitelist_path.

This slice is deliberately limited to the self-contained ban/whitelist surface:
every hoisted method touches ONLY members that move with it, so no dependent-base
qualification is needed and no consensus/sharechain-timing state is involved. The
bodies are relocated unchanged — pure de-duplication, behaviour identical.

The connection-coupled admin endpoints (admin_whitelist_add, admin_list_peers,
admin_drop_peer, admin_dial_peer) stay in NodeImpl for now: they reach into
BaseNode's connection maps and the outbound-dial members, so they belong to a
later slice that also hoists the outbound-connection manager.

DASH (still the 668-line skeleton, no ban/whitelist block yet) is untouched and
continues to derive from pool::BaseNode directly; it inherits this layer for free
once its node quick-wire lands and gives it the same block.

Net: -688 lines across the four coins, +243-line shared header (one source of
truth for four copies).

Regression evidence (this environment, system Boost/leveldb):
  - c2pool-{ltc,btc,bch,dgb,dash} all build green.
  - test_ltc_node: PASS.
  - core_test: 56/56 PASS (incl. version-gate transition suite).
  - ltc_pool_test harness constructs the node through the new base and enters the
    run loop with no crash.
