+++ src/impl/btc/node.hpp
@@ -8,6 +8,7 @@
 #include "messages.hpp"
 
 #include <pool/node.hpp>
+#include <pool/sharechain_node.hpp>
 #include <pool/protocol.hpp>
 #include <core/message.hpp>
 #include <core/reply_matcher.hpp>
@@ -32,8 +33,10 @@ struct ShareReplyData
     std::vector<chain::RawShare> m_raw_items;
 };
 
-class NodeImpl : public pool::BaseNode<btc::Config, btc::ShareChain, btc::Peer>
+class NodeImpl : public pool::SharechainNode<btc::Config, btc::ShareChain, btc::Peer>
 {
+
+    using base_t = pool::SharechainNode<btc::Config, btc::ShareChain, btc::Peer>;
     // Async share downloader:
     // ID = uint256 (matches sharereq id to sharereply id)
     // RESPONSE = parsed shares plus their original raw payloads
@@ -372,7 +375,6 @@ public:
     void set_max_peers(size_t count) { m_max_peers = count; }
 
     /// Set P2P ban duration (seconds).
-    void set_ban_duration(int seconds) { m_ban_duration = std::chrono::seconds(seconds); }
 
     /// Set cache size limits for memory control.
     void set_cache_limits(size_t max_shared, size_t max_known_txs, size_t max_raw) {
@@ -461,7 +463,6 @@ public:
     }
 
     /// Check whether a peer address is currently banned.
-    bool is_banned(const NetService& addr) const;
 
     /// ── Runtime admin API (pool peer bans + whitelist) ─────────────────
     /// All methods assumed to run on the io_context thread — callers
@@ -469,22 +470,15 @@ public:
     ///
     /// Returned JSON uses the shape:
     ///   {"ok": true|false, "error"?: "...", "bans": [...], "whitelist": [...]}
-    nlohmann::json admin_list_bans() const;
-    nlohmann::json admin_ban_ip(const std::string& ip, int duration_sec);
-    nlohmann::json admin_unban_ip(const std::string& ip);
-    nlohmann::json admin_list_whitelist() const;
     nlohmann::json admin_whitelist_add(const std::string& host, uint16_t port);
-    nlohmann::json admin_whitelist_remove(const std::string& host, uint16_t port);
     nlohmann::json admin_list_peers() const;
     nlohmann::json admin_drop_peer(const std::string& ip);
     nlohmann::json admin_dial_peer(const std::string& host, uint16_t port);
 
     /// Path to persisted whitelist file (~/.c2pool/pool_whitelist.json).
     /// Set by c2pool_refactored.cpp before start(); empty = no persistence.
-    void set_whitelist_path(const std::string& path);
 
     /// True if addr's IP matches a whitelist entry (IP or host:port).
-    bool is_whitelisted(const NetService& addr) const;
 
 protected:
     std::string m_software_version = "/c2pool:0.1/";  // overridden by set_software_version()
@@ -530,21 +524,13 @@ protected:
     std::set<NetService> m_outbound_addrs;     // successfully connected outbound peers
 
     // Peer banning: maps address → ban expiry time
-    std::map<NetService, std::chrono::steady_clock::time_point> m_ban_list;
-    std::chrono::seconds m_ban_duration{300}; // 5 minutes (configurable)
 
     // IP-only manual bans (admin endpoint). Keyed by IP string so the
     // operator can ban/unban without knowing the peer's source port.
-    std::map<std::string, std::chrono::steady_clock::time_point> m_ip_ban_list;
 
     // Whitelist: IPs that bypass is_banned() and host:port entries kept as
     // permanent dial targets. Persists across restart via m_whitelist_path.
-    std::set<std::string> m_whitelist_ips;
-    std::set<NetService> m_whitelist_hosts;
-    std::string m_whitelist_path;
 
-    void load_whitelist_from_disk();
-    void save_whitelist_to_disk() const;
 
     // Rate-limiting no longer needed: think() runs on a dedicated compute
     // thread (m_think_pool), serialized by m_think_running atomic flag.
